feat: Introduce Quick Revision mode
Add a new "revision" mode to the application, accessible from the main view. This mode will allow users to engage in a new type of learning experience, distinct from the existing solo and group quizzes.

This change also includes minor UI adjustments and scaling in quiz report and certificate views for better readability and presentation.